Update `one-page` example
