+++ examples/one-page/src/routes/+page.svelte
@@ -1,7 +1,7 @@
 <script>
-  import { t } from '$lib/translations'; 
+  import { t } from '$lib/translations';
 
-  const link = 'https://kit.svelte.dev'
+  const link = 'https://kit.svelte.dev';
 </script>
 
 <h1>{$t('content.title')}</h1>
